+++ src/theme/components/customCheckboxGroup.ts
@@ -1,0 +1,18 @@
+const CustomCheckboxGroup = {
+  parts: ["required", "helper"],
+  baseStyle: {
+    legend: {
+      marginBottom: 4,
+    },
+    required: {
+      marginLeft: 6,
+    },
+    helper: {
+      marginTop: 4,
+    },
+  },
+  sizes: {},
+  defaultProps: {},
+};
+
+export default CustomCheckboxGroup;